backend/requirements.txt
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.8.0
asgiref==3.7.2
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django>=4.2.0,<5.0.0
django-cors-headers>=4.3.1,<5.0.0
djangorestframework>=3.14.0,<4.0.0
drf-nested-routers==0.94.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai>=1.0.0
pydantic==2.10.6
pydantic_core==2.27.2
PyPDF2>=3.0.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
sqlparse>=0.4.4,<0.5.0
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
